% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynergyScreen.R
\docType{methods}
\name{boxplot,SynergyScreen-method}
\alias{boxplot,SynergyScreen-method}
\title{Box plot}
\usage{
\S4method{boxplot}{SynergyScreen}(x, what = "raw", ...)
}
\arguments{
\item{x}{object of class SynergyScreen}

\item{what}{type of data to plot, can be "raw" or "norm"}

\item{...}{additional arguments passed to the generic boxplot function}
}
\value{
See \code{graphics::\link[graphics]{boxplot}}
}
\description{
Produce box-and-whisker plot of a synergy screening experiment data
}
\examples{
data(sim15_screen3)
boxplot(sim15_screen3,"raw")
boxplot(sim15_screen3,"norm")

}
\seealso{
\code{BiocGenerics::\link[BiocGenerics]{boxplot}}, \code{graphics::\link[graphics]{boxplot}}, \code{\linkS4class{SynergyScreen}}
}

