% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRange.R
\docType{methods}
\name{checkRange}
\alias{checkRange}
\alias{checkRange,DRE-method}
\alias{checkRange,SynergyScreen-method}
\title{Check response ranges of dose-response experiments.}
\usage{
checkRange(object)

\S4method{checkRange}{DRE}(object)

\S4method{checkRange}{SynergyScreen}(object)
}
\arguments{
\item{object}{object of a supported class, e.g. DRE or SynergyScreen}
}
\value{
Object of the same class with @effect slot(s) populated with effect values and 
diagnostic message string(s) appended to the @comment slot(s)
}
\description{
A good titration should include doses that produce an informative range of effect values. This method checks if that is the case, prints 
diagnostic messages and populates @effect and @comment slots of the input object.  In case of a SynergyScreen object, these are inside 
individual members of its @dre_list slot
}
\details{
The assessment of a dose-response experiment's dose range is based on the range of produced effects.  

The effect is defined as fractional decrease in the response value, e.g. OD of the cell culture.  It ranges between 0 and 1, 
0 being untreated control and 1 complete growth inhibition and is defined as follows:

\emph{effect = 1 - response/control.response}

A dose-response experiment's dose range is considered "good" if all of the following is true
\itemize{
  \item the experiment includes observed effect values below 0.25 and above 0.75
  \item at least 30\% of all observed effect values fall between 0.25 and 0.75
}

The following diagnostics can be produced:
\itemize{
  \item "dose range is OK" - all criteria of good range specified above are met 
  \item "dose range too narrow" - the difference between the highest and the lowest effect values is < 0.5
  \item "poor coverage of mid-range effect region" - less than 30% of effect values fall between 0.25 and 0.75
  \item "high dose region not covered" - no effect values above 0.75
  \item "low dose region not covered" - no effect values below 0.25
}
}
\section{Methods (by class)}{
\itemize{
\item \code{DRE}: Assess dose range of a DRE object

\item \code{SynergyScreen}: Assess dose range of a SynergyScreen object

Assess dose ranges of all dose-response experiments included in a synergy screen, i.e. individual members of the @dre_list slot
of a SynergyScreen object.
}}
\examples{
data(sim15_screen3)

#  Check the ranges of all dose-response experiments in a screen
sim15_screen3 = checkRange(sim15_screen3)

#  checkRange populates @comment attributes of dre objects
comment(dre(sim15_screen3,"Cpd7"))

#  see all comments
lapply(dre_list(sim15_screen3), function(x) comment(x))

#  apply checkRange to an individual dose-response experiment
x = checkRange(dre(sim15_screen3,"Cpd1-Cpd2"))

}
\seealso{
\code{\linkS4class{DRE}}, \code{\linkS4class{SynergyScreen}}
}

