% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compound.R
\docType{class}
\name{Compound-class}
\alias{Compound-class}
\title{An S4 class to represent a compound}
\description{
An S4 class to represent a compound
}
\section{Slots}{

\describe{
\item{\code{name}}{compound name, character}

\item{\code{min.dose}}{minimum dose to be used in the dose-response curves, numeric}

\item{\code{max.dose}}{maximum dose to be used in the dose-response curves, numeric}

\item{\code{ic50}}{IC50 value, numeric}

\item{\code{m}}{slope parameter in the Hill's equation, numeric}
}}
\examples{
data(sim15_screen3)
compound(sim15_screen3,"Cpd1")

}

