% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompoundList.R
\docType{class}
\name{CompoundList-class}
\alias{CompoundList-class}
\title{An S4 class to represent a list of compounds}
\description{
Extends class list but requires all members to be of class Compound
}
\examples{
data(sim15_screen3)
compound_list(sim15_screen3)
  
}
\seealso{
\code{\linkS4class{Compound}}, \code{\link[base]{list}}
}

