% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSynergies.R
\docType{methods}
\name{computeSynergies}
\alias{computeSynergies}
\alias{computeSynergies,SynergyScreen-method}
\title{Compute synergy values}
\usage{
computeSynergies(object, ...)

\S4method{computeSynergies}{SynergyScreen}(object, effect = c(0.25, 0.5,
  0.75), model = "menls", control.weight = 10, span = 1, silent = T,
  ...)
}
\arguments{
\item{object}{object of class SynergyScreen}

\item{...}{other arguments passed to the model-fitting function, e.g. to loess or nls}

\item{effect}{numeric vector of effects, between 0 and 1}

\item{model}{character vector of model types.  Supported types include melm, menls and loess}

\item{control.weight}{weight to assing to the control.response, numeric}

\item{span}{span parameter for loess, numeric}

\item{silent}{logical: should warnings and error messages be suppressed?}
}
\value{
A SynergyScreen object with @synergies slot populated by a data frame of interaction index values
}
\description{
Synergy is measured as interaction index.  This function computes values of the interaction index for a set effective dose values for 
each synergy experiment in a screen.
}
\details{
See description of model types in \link{fit}
}
\section{Methods (by class)}{
\itemize{
\item \code{SynergyScreen}: Compute synergies in a screen
}}
\examples{
data(sim15_screen3)
sim15_screen3 = computeSynergies(sim15_screen3)
synergy_data(sim15_screen3)

}
\seealso{
\code{\linkS4class{SynergyScreen}}
}

