% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{designFile2SynergyScreen}
\alias{designFile2SynergyScreen}
\title{Initialize a SynergyScreen object from a design file}
\usage{
designFile2SynergyScreen(file)
}
\arguments{
\item{file}{char design file name}
}
\value{
a SynergyScreen object
}
\description{
Read in a design file and build a SynergyScreen object from it
}
\details{
This function creates a new SynergyScreen object and populates slots @design, @compound_list, @dre_list, and
}
\examples{
screen1 = designFile2SynergyScreen(system.file("extdata/8_cpds_1/8_compounds_1_design.csv", 
                                               package="SynergyScreen"))
str(screen1)

}
\seealso{
\code{\linkS4class{SynergyScreen}}, \code{\linkS4class{ScreenDesign}}
}

