% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRE-accessors.R
\docType{methods}
\name{DRE-accessors}
\alias{DRE-accessors}
\alias{comment}
\alias{comment,DRE-method}
\alias{comment<-}
\alias{comment<-,DRE-method}
\alias{compound.name}
\alias{compound.name,DRE-method}
\alias{compound.name<-}
\alias{compound.name<-,DRE-method}
\alias{control.response}
\alias{control.response,DRE-method}
\alias{control.response<-}
\alias{control.response<-,DRE-method}
\alias{dilution.factor,DRE-method}
\alias{dilution.factor<-,DRE-method}
\alias{dose}
\alias{dose,DRE-method}
\alias{dose<-}
\alias{dose<-,DRE-method}
\alias{effect}
\alias{effect,DRE-method}
\alias{effect<-}
\alias{effect<-,DRE-method}
\alias{fraction}
\alias{fraction,DRE-method}
\alias{fraction<-}
\alias{fraction<-,DRE-method}
\alias{ic25}
\alias{ic25,DRE-method}
\alias{ic25<-}
\alias{ic25<-,DRE-method}
\alias{ic50}
\alias{ic50,DRE-method}
\alias{ic50<-}
\alias{ic50<-,DRE-method}
\alias{ic75}
\alias{ic75,DRE-method}
\alias{ic75<-}
\alias{ic75<-,DRE-method}
\alias{model}
\alias{model,DRE-method}
\alias{model<-}
\alias{model<-,DRE-method}
\alias{name}
\alias{name,DRE-method}
\alias{name<-}
\alias{name<-,DRE-method}
\alias{response}
\alias{response,DRE-method}
\alias{response<-}
\alias{response<-,DRE-method}
\alias{true.ic50}
\alias{true.ic50,DRE-method}
\alias{true.ic50<-}
\alias{true.ic50<-,DRE-method}
\alias{true.m}
\alias{true.m,DRE-method}
\alias{true.m<-}
\alias{true.m<-,DRE-method}
\title{Accessor methods for class DRE}
\usage{
\S4method{name}{DRE}(object)

\S4method{compound.name}{DRE}(object)

\S4method{fraction}{DRE}(object)

\S4method{dose}{DRE}(object)

\S4method{response}{DRE}(object)

\S4method{control.response}{DRE}(object)

\S4method{effect}{DRE}(object)

\S4method{model}{DRE}(object)

\S4method{ic25}{DRE}(object)

\S4method{ic50}{DRE}(object)

\S4method{ic75}{DRE}(object)

\S4method{true.ic50}{DRE}(object)

\S4method{true.m}{DRE}(object)

\S4method{comment}{DRE}(object)

\S4method{dilution.factor}{DRE}(object)

\S4method{name}{DRE}(object) <- value

\S4method{compound.name}{DRE}(object) <- value

\S4method{fraction}{DRE}(object) <- value

\S4method{dose}{DRE}(object) <- value

\S4method{response}{DRE}(object) <- value

\S4method{control.response}{DRE}(object) <- value

\S4method{effect}{DRE}(object) <- value

\S4method{model}{DRE}(object) <- value

\S4method{ic25}{DRE}(object) <- value

\S4method{ic50}{DRE}(object) <- value

\S4method{ic75}{DRE}(object) <- value

\S4method{true.ic50}{DRE}(object) <- value

\S4method{true.m}{DRE}(object) <- value

\S4method{comment}{DRE}(object) <- value

\S4method{dilution.factor}{DRE}(object) <- value
}
\arguments{
\item{object}{object of class DRE}

\item{value}{replacement value for a slot}
}
\description{
Accessor methods for class DRE
}
\examples{
#  Retrieve a model and a comment for a dose-response experiment
data(sim15_screen3)
model(dre(sim15_screen3,"Cpd1"))
comment(dre(sim15_screen3,"Cpd1"))

#  Assign a new comment
x = dre(sim15_screen3,"Cpd1")
comment(x) = "test"
comment(x)

}

