% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreData.R
\docType{methods}
\name{dreData}
\alias{dreData}
\alias{dreData,DRE-method}
\alias{dreData,SynergyScreen-method}
\title{Retrive doses, normalized responses and effects}
\usage{
dreData(object, ...)

\S4method{dreData}{DRE}(object)

\S4method{dreData}{SynergyScreen}(object, names = NULL)
}
\arguments{
\item{object}{object of supported class, e.g. DRE or SynergyScreen}

\item{...}{method-specific arguments: see below}

\item{names}{the names of the dose-response experiments to retrieve data for, character}
}
\value{
data frame
}
\description{
Retrive doses, normalized responses and effects
}
\details{
\describe{
\item{DRE method}{returns a data frame is generated from @dose, @response and @effect slots}
\item{SynergyScreen method}{returns a data frame generated from @dose, @response and @effect slots of DRE objects contained 
in @dre_list slot of the SynergyScreen object.  Use parameter \code{names} to retrieve data for specific dose-response experiments.
If \code{names=NULL}, all dose-response data are retrieved}
}
}
\section{Methods (by class)}{
\itemize{
\item \code{DRE}: Retieve dose, normalized response and effect data from a DRE object

\item \code{SynergyScreen}: Retieve dose, normalized response and effect data from DRE objects contained in the @dre_list slot 
of a SynergyScreen object
}}
\examples{
data(sim15_screen3)
dreData(sim15_screen3)
dreData(dre(sim15_screen3,"Cpd1"))

}
\seealso{
\code{\linkS4class{DRE}}, \code{\linkS4class{SynergyScreen}}
}

