% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DREList.R
\docType{class}
\name{DREList-class}
\alias{DREList-class}
\title{An S4 class to represent a list of DRE objects}
\description{
Extends class list but requires all members to be of class DRE
}
\examples{
data(sim15_screen3)
dre_list(sim15_screen3)[1:3]

}
\seealso{
\code{\linkS4class{DRE}}, \code{\link[base]{list}}
}

