% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRModel.R
\docType{class}
\name{DRModel-class}
\alias{DRModel-class}
\title{A virtual S4 class representing a Dose-Response Model}
\description{
Extended by sub-classes for specific types of model
}
\examples{
data(sim15_screen3)
dre(sim15_screen3,"Cpd1")@model

}
\seealso{
\linkS4class{DRModelLOESS}, \linkS4class{DRModelMEnls}, \linkS4class{DRModelMElm}
}

