% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRModel.R
\docType{class}
\name{DRModelLOESS-class}
\alias{DRModelLOESS-class}
\title{An S4 class representing LOESS model of a dose-response curve}
\description{
An S4 class representing LOESS model of a dose-response curve
}
\section{Slots}{

\describe{
\item{\code{model}}{loess model object}

\item{\code{xlog}}{specifies if the x (dose) axis was log transformed before fitting a loess model}
}}
\examples{
data(sim15_screen3)

#  LOESS works
x = fit(dre(sim15_screen3,"Cpd10"),type="loess",span=1)
model(x)$loess
plot(x)

#  LOESS gives an odd result
x = fit(dre(sim15_screen3,"Cpd1"),type="loess",span=1)
model(x)$loess
plot(x)

}

