% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRModel.R
\docType{class}
\name{DRModelMElm-class}
\alias{DRModelMElm-class}
\title{An S4 class representing median effect model of a dose-response curve fitted using lm}
\description{
An S4 class representing median effect model of a dose-response curve fitted using lm
}
\section{Slots}{

\describe{
\item{\code{model}}{object of class \code{\link[stats]{lm}}}

\item{\code{control.response}}{response value of untreated control, numeric}
}}
\examples{
data(sim15_screen3)
x = fit(dre(sim15_screen3,"Cpd10"),type="melm")
model(x)$melm
plot(x)

}

