% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRModel.R
\docType{class}
\name{DRModelMEnls-class}
\alias{DRModelMEnls-class}
\title{An S4 class representing median effect model of a dose-response curve fitted using nls}
\description{
An S4 class representing median effect model of a dose-response curve fitted using nls
}
\section{Slots}{

\describe{
\item{\code{model}}{object of class \code{\link[stats]{nls}}}
}}
\examples{
data(sim15_screen3)
model(dre(sim15_screen3, "Cpd1"))
plot(dre(sim15_screen3, "Cpd1"))

}

