% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{effectiveDose}
\alias{effectiveDose}
\title{Compute the dose that causes specified fractional effect}
\usage{
effectiveDose(dre, effect, model = "menls")
}
\arguments{
\item{dre}{a DRE object}

\item{effect}{fractional effect value, numeric}

\item{model}{name of the model to use, character}
}
\value{
dose value that causes the effect, numeric
}
\description{
Take a DRE (dose-response experiment) object and compute the dose that causes the specified fractional effect
}
\details{
effect value should be between 0 and 1, e.g. 0.5 to compute interaction index at IC50, 0.25 to compute IC25 (25% inhibition) etc.
Corresponding value of the response variable is defined as control.response*(1-effect)

If the effect is not achieved by doses between 0 and max dose for the dose-response experiment,
effectiveDose produces a warning and returns NA
}
\examples{
data(sim15_screen3)
effectiveDose(dre(sim15_screen3,"Cpd1"),0.5)

}
\seealso{
\code{\linkS4class{DRE}}
}

