% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSynergies.R
\docType{methods}
\name{findSynergies}
\alias{findSynergies}
\alias{findSynergies,SynergyScreen-method}
\title{Find pairs of compounds that show synergy or antagonism}
\usage{
findSynergies(object, ...)

\S4method{findSynergies}{SynergyScreen}(object, statistic, threshold, direction,
  model = NULL)
}
\arguments{
\item{object}{object of class SynergyScreen}

\item{...}{method-specific arguments: see below}

\item{statistic}{type of statistic to use to find synergistic or antagonistic compound pairs.  Can be "min", "max", "mean" or "median".}

\item{threshold}{threshold value that needs to be surpassed by the statistic}

\item{direction}{can be "higher" or "lower".  See Details.}

\item{model}{character.  Only consider results from the specified model.  If NULL, consider all models.}
}
\value{
object of type SynergyData
}
\description{
Find pairs of compounds that show synergy or antagonism
}
\details{
This function filters the output of synergy_data(object) to report only compounds that are synergistic or antagonistic.
This is done by computing a statistic of the set of interaction index values for each pair of compounds and comparing it to 
the threshold.  

\strong{Example 1}.  To find pairs of compounds that show consistent synergy for all analyzed effect sizes, 
one may use statistic = "max", threshold = 0.7 and direction = "lower".  This will select pairs of compounds whose interaction
index does not exceed 0.7.

\strong{Example 2}.  To find pairs of compounds that show strong synergy that may or may not be consistent for all effect sizes,
one may use statistic = "min", threshold = 0.3 and direction = "lower".  This will select pairs of compounds whose interaction
index is below 0.3 for at least one effect size.
}
\section{Methods (by class)}{
\itemize{
\item \code{SynergyScreen}: Find pairs of compounds that show synergy or antagonism
}}
\examples{
data(sim15_screen3)

#  Find strong synergies
findSynergies(sim15_screen3, statistic="max", threshold=0.5, direction="lower")

#  Find strong antagonisms
findSynergies(sim15_screen3, statistic="min", threshold=2, direction="higher")

}
\seealso{
\code{\linkS4class{SynergyScreen}}, \code{\link[=SynergyScreen-accessors]{synergy_data}}
}

