% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDesign.R
\docType{methods}
\name{generateDesign}
\alias{generateDesign}
\alias{generateDesign,DREList-method}
\alias{generateDesign,SynergyScreen-method}
\title{Generate design of microtiter plate-based synergy screening experiments.}
\usage{
generateDesign(object, ...)

\S4method{generateDesign}{DREList}(object, dre.orientation, control.orientation,
  use.edge, digits = 3)

\S4method{generateDesign}{SynergyScreen}(object, type = "5 doses exp",
  dre.orientation = NULL, control.orientation = NULL,
  dose.series.type = NULL, use.edge = NULL, singles = T, pairs = T,
  digits = 3)
}
\arguments{
\item{object}{an object of supported type, e.g. \code{DREList} or \code{SynergyScreen}}

\item{...}{method-specific arguments: see below}

\item{dre.orientation}{character.  Must be set if \code{type}="custom". 
Specifies if dose-response experiments are to be in rows or columns.
Supported values are "rows" and "columns"}

\item{control.orientation}{character.  Must be set if \code{type}="custom". 
Specifies if untreated controls are to be in rows or columns.
Supported values are "rows" and "columns"}

\item{use.edge}{logical.  Must be set if \code{type}="custom".
Specifies if wells at the edge of the plate, e.g. rows "A" and "H" and columns "1" and "12", 
are to be used for dose series and untreated controls or left blank.}

\item{digits}{number of decimal places in dose values}

\item{type}{design type.  Supported types include "5 doses exp", "8 doses linear", "12 doses exp", 
and "custom".  See Details.}

\item{dose.series.type}{character type of dose series.  Must be set if \code{type}="custom".
Supported values are "linear" and "exp".}

\item{singles}{logical: if singles = T, include dose-response experiments for individual compounds 
(\code{SynergyScreen} method)}

\item{pairs}{logical: if pairs = T, include dose-response experiments for compound pairs 
(\code{SynergyScreen} method)}
}
\value{
\describe{
\item{SynergyScreen method}{returns a SynergyScreen object with populated @design, @dre_list and, 
if \code{pairs=T}, @synergy_experiment_list slots} 
\item{DREList method}{returns a data frame that represents experimental design}
}
}
\description{
Design specifies compound treatments, untreated controls and blanks that go in each well of a set of 
microtiter plates.
}
\details{
Screen design is represented as a data frame.  A \code{\linkS4class{SynergyScreen}} object has slots
}
\section{Methods (by class)}{
\itemize{
\item \code{DREList}: Generate design from a \code{DREList} object 

Taking a \code{DREList} object as input, generate a microtiter plate design showing what to put in 
each well.

\item \code{SynergyScreen}: Taking a \code{SynergyScreen} object as input, generate a microtiter plate
 design showing what to put in each well.

This method also sets up dose-response experiment objects in object@dre_list 
and synergy experiment objects in object@synergy_experiment_list
}}
\examples{
compounds = readCompoundFile(system.file("extdata/8_cpds_1/8_compounds_1.csv",
                                         package="SynergyScreen"))
screen = new("SynergyScreen", compound_list=compounds)
screen = generateDesign(screen, type="8 doses linear")
design(screen)

}
\seealso{
\code{\linkS4class{SynergyScreen}}, \code{\linkS4class{DREList}}, \code{\linkS4class{DRE}}, 
\code{\linkS4class{SynergyExperimentList}}, \code{\linkS4class{SynergyExperiment}}
}

