% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynergyExperiment.R
\docType{methods}
\name{initialize,SynergyExperiment-method}
\alias{initialize,SynergyExperiment-method}
\title{Initialise a SynergyExperiment object}
\usage{
\S4method{initialize}{SynergyExperiment}(.Object, name = NULL,
  mixture.dre.name, compound.dre.name, dre.list)
}
\arguments{
\item{.Object}{object of class SynergyExperiment}

\item{name}{optional name of the experiment, character}

\item{mixture.dre.name}{name of the mixture dose-response experiment included in the synergy experiment, character}

\item{compound.dre.name}{names of individual compound dose-response experiments that are included in the synergy experiment, character}

\item{dre.list}{object of class DREList or list of DRE objects}
}
\value{
SynergyExperiment object
}
\description{
Create a new SynergyExperiment object from appropriate DRE objects,
making sure they form a coherent experiment
}
\details{
A single ray experiment is assumed in the current implementation, i.e. there should be only one mixture
dose-response experiment.  There should also be 2 or more compound dose-response experiments.

The dre.list argument is used to check that compound.dre.name and mixture.dre.name
refer to valid dose-response experiments and make sure compounds are listed in the same order in the synergy experiment object as 
in the corresponding mixture dose-response experiment.
}
\examples{
data(sim15_screen3)
new("SynergyExperiment", name="Cpd1-Cpd2", mixture.dre.name="Cpd1-Cpd2", 
    compound.dre.name=c("Cpd1","Cpd2"), dre_list(sim15_screen3))

}
\seealso{
\code{\linkS4class{DRE}}
}

