% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{interactionIndex}
\alias{interactionIndex}
\title{Compute Interaction Index based on Loewe Additivity}
\usage{
interactionIndex(experiment, effect, dre.list, model = "menls")
}
\arguments{
\item{experiment}{an object of class SynergyExperiment}

\item{effect}{fractional effect value, numeric}

\item{dre.list}{a list of DRE objects}

\item{model}{name of the model to use (the same model must be present in all dre.list members)}
}
\value{
interaction index value, numeric
}
\description{
Compute Interaction Index from SynergyExperiment and the matching DRE (dose-response experiment) objects
}
\details{
effect value should be between 0 and 1, e.g. 0.5 to compute interaction index at IC50, 0.25 to compute IC25 (25% inhibition) etc.
Corresponding value of the response variable is defined as control.response*(1-effect)
}
\examples{
data(sim15_screen3)
interactionIndex(synergy_experiment(sim15_screen3,"Cpd1-Cpd2"), 0.5, dre_list(sim15_screen3))

}
\seealso{
\code{\linkS4class{SynergyExperiment}}, \code{\linkS4class{DRE}}
}

