% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelledData.R
\docType{methods}
\name{labelledData}
\alias{labelledData}
\alias{labelledData,SynergyScreen-method}
\title{Retrieve labelled data}
\usage{
labelledData(object, what)

\S4method{labelledData}{SynergyScreen}(object, what = "raw")
}
\arguments{
\item{object}{object of class SynergyScreen}

\item{what}{type of data to retrieve, can be "raw" or "norm"}
}
\value{
data frame containing full screen design information along with response data
}
\description{
Retrieve synergy screen data as a data frame containing full screen design information along with response data
}
\section{Methods (by class)}{
\itemize{
\item \code{SynergyScreen}: Retrieve labelled data
}}
\examples{
data(sim15_screen3)
head(labelledData(sim15_screen3, what="raw"),n=30)
 
}
\seealso{
\code{\linkS4class{SynergyScreen}}
}

