% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{meq}
\alias{meq}
\title{Median Effect Equation}
\usage{
meq(d, ic50, m, control.response)
}
\arguments{
\item{d}{dose}

\item{ic50}{IC50 parameter of the median effect equation}

\item{m}{m parameter of the median effect equation}

\item{control.response}{response value of untreated control}
}
\value{
value of the Median Effect Equation for dose d
}
\description{
Compute the value of the Median Effect Equation for dose d
}
\examples{
meq(d=seq(0,20,by=5), ic50=2, m=1, control.response=1)

}

