% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynergyScreen.R
\docType{methods}
\name{normalize,SynergyScreen-method}
\alias{normalize,SynergyScreen-method}
\title{Normalize synergy screen data}
\usage{
\S4method{normalize}{SynergyScreen}(object, floor = 0.001)
}
\arguments{
\item{object}{object of class SynergyScreen}

\item{floor}{numeric, normalized non-blank response values smaller than this number are set to this number}
}
\value{
object of class SynergyScreen with populated @norm_data slot and @response and @control.response 
slots of @dre_list members
}
\description{
Normalization corrects for plate bias and subtracts blank response
}
\details{
Normalization assumes that there is a multiplicative plate bias.  A bias factor is computed for each plate based on  
untreated control response values.  A plate's bias is its mean divided by the global mean.  To normalize, all response values 
in a plate are divided by the plate's bias factor.
}
\examples{
# Initialize a screen
cpds <- readCompoundFile(system.file("extdata/15_cpds_simulation/compounds_1.csv",
                                    package="SynergyScreen"))
screen1 <- new("SynergyScreen",compound_list=cpds)
screen1 <- generateDesign(screen1,pairs=FALSE)

#  Simulate compound growth curve characteristics and raw dose-response data
set.seed(20141120)
screen1 <- simulate(screen1)
boxplot(screen1,"raw")

#  Normalize screen data, adjusting for plate bias
screen1 <- normalize(screen1)
boxplot(screen1,"norm")

}
\seealso{
\code{\linkS4class{SynergyScreen}}, \code{BiocGenerics::\link[BiocGenerics]{normalize}}
}

