% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRE.R
\docType{methods}
\name{plot,DRE,missing-method}
\alias{plot,DRE,missing-method}
\title{Plot a DRE object}
\usage{
\S4method{plot}{DRE,missing}(x, y, linearize = F, xlab = NULL,
  ylab = NULL, log = NULL, ...)
}
\arguments{
\item{x}{object of class DRE}

\item{y}{no parameter \code{y} is defined for this method}

\item{linearize}{logical, defaults to FALSE (see Details)}

\item{xlab}{a title for the x axis.  If left unspecified, appropriate defaults are chosen for sigmoid and linearized plots.}

\item{ylab}{a title for the y axis.  If left unspecified, appropriate defaults are chosen for sigmoid and linearized plots.}

\item{log}{a character string which contains "x" if the x axis is to be logarithmic, "y" if the y axis is to be logarithmic 
and "xy" or "yx" if both axes are to be logarithmic.  If left unspecified, appropriate defaults are chosen for sigmoid and linearized plots.}

\item{...}{additional arguments passed to the generic \code{\link[graphics]{plot}} function}
}
\description{
Plot dose-response data and, if known, a fitted curve and/or a "true" curve on the basis of which the data have been simulated
}
\details{
If "linearize" is set to FALSE (default), plot response vs. dose.  
If TRUE, plot log(fraction.affected/fraction.unaffected) vs. log(dose), which is linear if the dose-response relationship follows 
the median effect law
}
\examples{
data(sim15_screen3)
plot(dre(sim15_screen3,"Cpd1"))

}
\seealso{
\code{\linkS4class{DRE}}, \code{\link[graphics]{plot}}
}

