% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynergyExperiment.R
\docType{methods}
\name{plot,SynergyExperiment,missing-method}
\alias{plot,SynergyExperiment,missing-method}
\title{Plot a SynergyExperiment object}
\usage{
\S4method{plot}{SynergyExperiment,missing}(x, y, dre.list, effect = c(0.25,
  0.5, 0.75), model = "menls", xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{object of class SynergyExperiment}

\item{y}{no parameter \code{y} is defined for this method}

\item{dre.list}{a list of DRE objects where the data come from}

\item{effect}{vector of fractional effect values to draw the isoboles for, numeric}

\item{model}{name of the growth curve model to use, character}

\item{xlab}{a title for the x axis.  If left unspecified, appropriate default is chosen.}

\item{ylab}{a title for the y axis.  If left unspecified, appropriate default is chosen.}

\item{...}{additional arguments passed to the generic \code{\link[graphics]{plot}} function}
}
\description{
Isobole plot of a single ray experiment
}
\details{
\link{plotIsoboles} is a convenient wrapper for this method.

effect value should be between 0 and 1, e.g. 0.5 to compute interaction index at IC50, 0.25 to compute IC25 (25% inhibition) etc.
Corresponding value of the response variable is defined as control.response*(1-effect)
}
\examples{
data(sim15_screen3)
plot(synergy_experiment(sim15_screen3,"Cpd1-Cpd2"), dre.list=dre_list(sim15_screen3))

}
\seealso{
\link{plotIsoboles} \code{\linkS4class{SynergyExperiment}}, \code{\linkS4class{DRE}}, \code{\link{effectiveDose}}, \code{\link[graphics]{plot}}
}

