% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynergyScreen.R
\docType{methods}
\name{plot,SynergyScreen,missing-method}
\alias{plot,SynergyScreen,missing-method}
\title{Plot a SynergyScreen object.}
\usage{
\S4method{plot}{SynergyScreen,missing}(x, y, effect = 0.5, model = "menls",
  ...)
}
\arguments{
\item{x}{a SynergyScreen object}

\item{y}{no parameter \code{y} is defined for this method}

\item{effect}{numeric, fractional effect value to plot the matrix for}

\item{model}{name of the growth curve model to use, character}

\item{...}{additional arguments passed to the \code{\link[stats]{heatmap}} function}
}
\description{
Plot matrix of pairwise interaction index values for a synergy screen.
}
\details{
The matrix visualizes log2(interaction index) values for the specified effect size.  
The values are visualized on a red-yellow-blue color scale, red indicating synergy,
yellow no interaction and blue antagonism.

The index is visualized on log scale because it is a sum of two ratios.  
Distributions of ratios are highly assymetrical, e.g. 0 to infinity with a mode around 1
Log transform is commonly used to normalize them.
}
\examples{
data(sim15_screen3)
plot(sim15_screen3)

}
\seealso{
\code{\linkS4class{SynergyScreen}}, \code{\link[stats]{heatmap}}
}

