% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIsoboles.R
\docType{methods}
\name{plotIsoboles}
\alias{plotIsoboles}
\alias{plotIsoboles,SynergyScreen-method}
\title{Plot isoboles for a specified synergy experiment}
\usage{
plotIsoboles(object, ...)

\S4method{plotIsoboles}{SynergyScreen}(object, experiment, effect = c(0.25,
  0.5, 0.75), model = "menls", xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{object}{a SynergyScreen object}

\item{...}{additional arguments passed to the generic \code{\link[graphics]{plot}} function}

\item{experiment}{string, name of the synergy experiment to plot}

\item{effect}{vector of fractional effect values to draw the isoboles for, numeric}

\item{model}{name of the growth curve model to use, character}

\item{xlab}{a title for the x axis.  If left unspecified, appropriate default is chosen.}

\item{ylab}{a title for the y axis.  If left unspecified, appropriate default is chosen.}
}
\description{
Plot isoboles for a synergy experiment that is part of a screen.
}
\details{
The experiment must be listed in the synergy_experiment_list slot of a SynergyScreen object

The effect values should be between 0 and 1, e.g. 0.5 to compute interaction index at IC50, 0.25 to compute IC25 (25% inhibition) etc.
Corresponding value of the response variable is defined as control.response*(1-effect)
}
\examples{
data(sim15_screen3)
plotIsoboles(sim15_screen3,"Cpd1-Cpd2")

}
\seealso{
\code{\link[=plot,SynergyExperiment,missing-method]{plot,SynergyExperiment-method}}, \code{\linkS4class{SynergyScreen}}
}

