% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{plotMEQ}
\alias{plotMEQ}
\title{Plot median effect equation}
\usage{
plotMEQ(ic50, m, control.response, dose.range)
}
\arguments{
\item{ic50}{IC50 parameter of the median effect equation}

\item{m}{m parameter of the median effect equation}

\item{control.response}{response value of untreated control}

\item{dose.range}{range of doses, a vector of 2 real positive numbers}
}
\description{
Plot median effect equation
}
\examples{
plotMEQ(ic50=2, m=1, control.response=1, dose.range=c(0.1,30))

}

