% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict,DRModelLOESS-method}
\alias{predict,DRModelLOESS-method}
\title{Predict response values from a DRModelLOESS object}
\usage{
\S4method{predict}{DRModelLOESS}(object, dose, ...)
}
\arguments{
\item{object}{object of class DRModelLOESS}

\item{dose}{a vector of dose values for which to compute response, numeric}

\item{...}{other arguments to be passed to \code{\link[stats]{predict.loess}}}
}
\value{
numeric vector of response values
}
\description{
Given a vector of dose values, computes response on linear scale
}
\details{
A DRModelLOESS object contains a loess model for log(response) vs. dose.  
This method Executes \code{\link[stats]{predict.loess}} and exponentiates to return response on linear scale
}
\examples{
data(sim15_screen3)
x = fit(dre(sim15_screen3,"Cpd10"),type="loess")
model(x)$loess
plot(x)
dose(x)
y = predict(model(x)$loess, dose=seq(0.4,4,by=0.4))
data.frame(dose = seq(0.4,4,by=0.4), y)


}

