% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict,DRModelMElm-method}
\alias{predict,DRModelMElm-method}
\title{Predict response values from a DRModelMElm object}
\usage{
\S4method{predict}{DRModelMElm}(object, dose, ...)
}
\arguments{
\item{object}{object of class DRModelMElm}

\item{dose}{a vector of dose values for which to compute response, numeric}

\item{...}{other arguments to be passed to \code{\link[stats]{predict.lm}}}
}
\value{
numeric vector of response values
}
\description{
Given a vector of dose values, computes response on linear scale
}
\details{
A DRModelMElm object contains a lm model for log(response) vs. dose.  
This method Executes \code{\link[stats]{predict.lm}} and exponentiates to return response on linear scale
}
\examples{
data(sim15_screen3)
x = fit(dre(sim15_screen3,"Cpd10"),type="melm")
model(x)$melm
plot(x)
dose(x)
y = predict(model(x)$melm, dose=seq(0.4,4,by=0.4))
data.frame(dose = seq(0.4,4,by=0.4), y)

}

