% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict,DRModelMEnls-method}
\alias{predict,DRModelMEnls-method}
\title{Predict response values from a DRModelMEnls object}
\usage{
\S4method{predict}{DRModelMEnls}(object, dose, ...)
}
\arguments{
\item{object}{object of class DRModelMEnls}

\item{dose}{a vector of dose values for which to compute response, numeric}

\item{...}{other arguments to be passed to \code{\link[stats]{predict.nls}}}
}
\value{
numeric vector of response values
}
\description{
Given a vector of dose values, computes response on linear scale
}
\details{
A DRModelMEnls object contains a nls model for log(response) vs. dose.  
This method Executes \code{\link[stats]{predict.nls}} and exponentiates to return response on linear scale
}
\examples{
data(sim15_screen3)
y = predict(model(dre(sim15_screen3, "Cpd10"))$menls, dose=seq(0.4,4,by=0.4))
data.frame(dose = seq(0.4,4,by=0.4), y)

}

