% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{readCompoundFile}
\alias{readCompoundFile}
\title{Load a spreadsheet of compounds}
\usage{
readCompoundFile(file)
}
\arguments{
\item{file}{file to read from}
}
\value{
a list of Compound objects
}
\description{
Reads in a CSV file with columns "compound" and "max.dose" into a list of Compound objects
}
\examples{
cpds = readCompoundFile(system.file("extdata/15_cpds_simulation/compounds_1.csv",
                                    package="SynergyScreen"))
cpds[[1]]

}

