% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{readTecanData}
\alias{readTecanData}
\title{Extract raw data from a set of Tecan files}
\usage{
readTecanData(path, time)
}
\arguments{
\item{path}{path to a folder with Tecan files, character.  The folder must contain Tecan files in CSV format only.}

\item{time}{time point, hours, numeric}
}
\description{
Parse Tecan files one by one and assemble a data frame of results.  
This data frame is suitable for conversion to a SynergyData object.
}
\seealso{
\linkS4class{SynergyData}
}

