% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{readTecanFile}
\alias{readTecanFile}
\title{Read Tecan file}
\usage{
readTecanFile(file, time)
}
\arguments{
\item{file}{Tecan file name, character}

\item{time}{time point, hours, numeric}
}
\value{
Data frame of response values in wells.  It is suitable for conversion to a SynergyData object upon addition of a 
\code{plate} column.
}
\description{
Read in a Tecan file in CSV format and extract the difference in response values (e.g. OD) at a specified time point vs. zero time
}
\seealso{
\link{readTecanData}
}

