% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{readWideData}
\alias{readWideData}
\title{Read synergy screen data in a wide format}
\usage{
readWideData(file)
}
\arguments{
\item{file}{character file}
}
\value{
data frame suitable for conversion to a ScreenData object
}
\description{
Read in synergy screen data in a wide format and reshape it into a long format
}
\seealso{
\code{\linkS4class{ScreenData}}
}

