% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{readWideDataFlex}
\alias{readWideDataFlex}
\title{Read synergy screen data in a wide format from a CSV file of arbitrary layout}
\usage{
readWideDataFlex(file, row, col)
}
\arguments{
\item{file}{character file}

\item{row}{numeric row coordinates of plate data blocks}

\item{col}{character column coordinates of plate data blocks}
}
\value{
data frame suitable for conversion to a ScreenData object
}
\description{
Read in synergy screen data in a wide format and reshape it into a long format
}
\details{
This function reads data from a CSV file.  Data for each plate are in 8 row X 12 column blocks.
Parameters \code{row} and \code{col} specify the coordinates of the upper left corner of each block.
\code{col} is type character, because the function expects Excel column names.
}
\seealso{
\code{\linkS4class{ScreenData}}
}

