% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenData.R
\docType{class}
\name{ScreenData-class}
\alias{ScreenData-class}
\title{S4 class for a synergy screen data}
\description{
A ScreenData class object contains a special data frame containing response values for microtiter plate wells
}
\details{
Initial implementation supports 96-well, 8*12 plates.  
Plate rows should be named A-H and columns 1-12.  ScreenData data frame must contain the following columns:
\enumerate{
  \item \code{plate} plate identifier
  \item \code{column} column number, numeric, integer between 1 and 12
  \item \code{row} row, character or factor with values or levels A-H
  \item \code{response} numeric response value, e.g. OD
}
Additional columns are allowed and will be ignored by SynergyScreen software.
Correctness of the data frame structure is verified by a validity method and can be checked by calling validObject().
}
\examples{
data(sim15_screen3)
head(raw_data(sim15_screen3),30)
str(raw_data(sim15_screen3))

}
\seealso{
\code{\link[base]{data.frame}}, \code{\link[methods]{validObject}}
}

