% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenDesign.R
\docType{class}
\name{ScreenDesign-class}
\alias{ScreenDesign-class}
\title{S4 class for a synergy screen design}
\description{
A ScreenDesign class object contains a special data frame specifying design of a screen, i.e. what goes into each well
of a set of plates.
}
\details{
Initial implementation supports 96-well, 8*12 plates and mixtures of 2 compounds.  
Plate rows should be named A-H and columns 1-12.  Design data frame must contain the following columns:
\enumerate{
  \item \code{plate} plate identifier
  \item \code{column} column number, numeric, integer between 1 and 12
  \item \code{row} row, character or factor with values or levels A-H
  \item \code{experiment} experiment name, character or factor
  \item \code{cpd1} compound 1 name, character or factor
  \item \code{dose1} compound 1 dose, numeric
  \item \code{cpd2} compound 2 name, character or factor
  \item \code{dose2} compound 2 dose, numeric
}
Additional columns are allowed and will be ignored by SynergyScreen software.
Correctness of the data frame structure is verified by a validity method and can be checked by calling validObject().
}
\examples{
data(sim15_screen3)
head(design(sim15_screen3),n=30)  
str(design(sim15_screen3))

}
\seealso{
\code{\link[base]{data.frame}}, \code{\link[methods]{validObject}}
}

