% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynergyData.R
\docType{class}
\name{SynergyData-class}
\alias{SynergyData-class}
\title{S4 class for a synergy data}
\description{
A SynergyData class object contains a special data frame containing interaction index values from a synergy screen
}
\details{
SynergyData data frame must contain the following columns:
\enumerate{
  \item \code{experiment} the name of a synergy experiment, character or factor
  \item \code{model} the name of the model used to fit dose-resoponse curves
  \item \code{effect} effect value, numeric between 0 and 1
  \item \code{interaction.index} interaction index value, a positive number
}
Additional columns are allowed and will be ignored by SynergyScreen software.
Correctness of the data frame structure is verified by a validity method and can be checked by calling validObject().
}
\examples{
data(sim15_screen3)
head(synergy_data(sim15_screen3))
str(synergy_data(sim15_screen3))

}
\seealso{
\code{\link[base]{data.frame}}, \code{\link[methods]{validObject}}, \code{\linkS4class{SynergyScreen}}
}

