% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynergyExperiment.R
\docType{class}
\name{SynergyExperiment-class}
\alias{SynergyExperiment-class}
\title{S4 class for a synergy experiment}
\description{
A SynergyExperiment class object contains DRE objects for individual compounds 
and fixed-ratio mixture ray(s)
}
\section{Slots}{

\describe{
\item{\code{name}}{of the experiment}

\item{\code{mixture.dre.name}}{a vector of dose-response experiment names of mixtures, character}

\item{\code{compound.dre.name}}{a vector of dose-response experiment names of individual compounds, character}
}}
\examples{
data(sim15_screen3)
synergy_experiment(sim15_screen3,"Cpd1-Cpd2")

}
\seealso{
\code{\linkS4class{DRE}}
}

