% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynergyExperimentList.R
\docType{class}
\name{SynergyExperimentList-class}
\alias{SynergyExperimentList-class}
\title{An S4 class to represent a list of SynergyExperiment objects}
\description{
Extends class list but requires all members to be of class SynergyExperiment
}
\examples{
data(sim15_screen3)
synergy_experiment_list(sim15_screen3)[[1]]

}
\seealso{
\code{\linkS4class{SynergyExperiment}}, \code{\link[base]{list}}
}

