% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynergyScreen-accessors.R
\docType{methods}
\name{SynergyScreen-accessors}
\alias{SynergyScreen-accessors}
\alias{compound}
\alias{compound,SynergyScreen-method}
\alias{compound_list}
\alias{compound_list,SynergyScreen-method}
\alias{compound_list<-}
\alias{compound_list<-,SynergyScreen-method}
\alias{design,SynergyScreen-method}
\alias{design<-,SynergyScreen-method}
\alias{dre}
\alias{dre,SynergyScreen-method}
\alias{dre_list}
\alias{dre_list,SynergyScreen-method}
\alias{dre_list<-}
\alias{dre_list<-,SynergyScreen-method}
\alias{norm_data}
\alias{norm_data,SynergyScreen-method}
\alias{norm_data<-}
\alias{norm_data<-,SynergyScreen-method}
\alias{raw_data}
\alias{raw_data,SynergyScreen-method}
\alias{raw_data<-}
\alias{raw_data<-,SynergyScreen-method}
\alias{synergy_data}
\alias{synergy_data,SynergyScreen-method}
\alias{synergy_data<-}
\alias{synergy_data<-,SynergyScreen-method}
\alias{synergy_experiment}
\alias{synergy_experiment,SynergyScreen-method}
\alias{synergy_experiment_list}
\alias{synergy_experiment_list,SynergyScreen-method}
\alias{synergy_experiment_list<-}
\alias{synergy_experiment_list<-,SynergyScreen-method}
\title{Accessor methods for class SynergyScreen}
\usage{
norm_data(object)

norm_data(object) <- value

raw_data(object)

raw_data(object) <- value

synergy_data(object)

synergy_data(object) <- value

compound_list(object)

compound_list(object) <- value

compound(object, name)

dre_list(object)

dre_list(object) <- value

dre(object, name)

synergy_experiment_list(object)

synergy_experiment_list(object) <- value

synergy_experiment(object, name)

\S4method{design}{SynergyScreen}(object)

\S4method{design}{SynergyScreen}(object) <- value

\S4method{norm_data}{SynergyScreen}(object)

\S4method{norm_data}{SynergyScreen}(object) <- value

\S4method{raw_data}{SynergyScreen}(object)

\S4method{raw_data}{SynergyScreen}(object) <- value

\S4method{synergy_data}{SynergyScreen}(object)

\S4method{synergy_data}{SynergyScreen}(object) <- value

\S4method{compound_list}{SynergyScreen}(object)

\S4method{compound_list}{SynergyScreen}(object) <- value

\S4method{compound}{SynergyScreen}(object, name)

\S4method{dre_list}{SynergyScreen}(object)

\S4method{dre_list}{SynergyScreen}(object) <- value

\S4method{dre}{SynergyScreen}(object, name)

\S4method{synergy_experiment_list}{SynergyScreen}(object)

\S4method{synergy_experiment_list}{SynergyScreen}(object) <- value

\S4method{synergy_experiment}{SynergyScreen}(object, name)
}
\arguments{
\item{object}{object of class SynergyScreen}

\item{value}{replacement value for a slot.  See Details}

\item{name}{name of a list member to access, e.g. specific dose-response experiment or synergy experiment}
}
\description{
Accessor methods for class SynergyScreen
}
\details{
Details on specific methods
}
\section{Replacement method for @norm_data (norm_data<-)}{

Replacement value for @norm_data can be a \linkS4class{ScreenData} object or a data frame

This method also sets @response and @control.response slots of @dre_list members.
}

\section{Replacement method for @raw_data (raw_data<-)}{

Replacement value for @norm_data can be a \linkS4class{ScreenData} object or a data frame
}

\section{Replacement method for @synergy_data (synergy_data<-)}{

Replacement value for @synergy_data can be a \linkS4class{SynergyData} object or a data frame
}
\examples{
data(sim15_screen3)
head(norm_data(sim15_screen3))
dre(sim15_screen3,"Cpd1")

}
\seealso{
\linkS4class{SynergyScreen}, \link{dreData}, \link{labelledData}
}

