% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynergyScreen.R
\docType{class}
\name{SynergyScreen-class}
\alias{SynergyScreen-class}
\title{S4 class for a synergy screen}
\description{
A SynergyScreen class object contains all data for an entire synergy screen
}
\section{Slots}{

\describe{
\item{\code{design}}{design of the screen, object of class ScreenDesign}

\item{\code{raw_data}}{raw screen data, object of class ScreenData}

\item{\code{norm_data}}{normalized screen data, object of class ScreenData}

\item{\code{compound_list}}{list of compounds screened, object of class CompoundList}

\item{\code{dre_list}}{list of dose-response experiments, object of class DREList}

\item{\code{synergy_experiment_list}}{list of synergy experiments, object of class SynergyExperimentList}

\item{\code{synergy_data}}{interaction index data}
}}
\examples{
data(sim15_screen3)
head(design(sim15_screen3),n=30)
head(raw_data(sim15_screen3))
head(norm_data(sim15_screen3))
compound_list(sim15_screen3)[1:2]
dre_list(sim15_screen3)[1:2]
synergy_experiment_list(sim15_screen3)[1:2]
head(synergy_data(sim15_screen3))

}
\seealso{
\code{\linkS4class{ScreenDesign}}, \code{\linkS4class{ScreenData}}, 
\code{\linkS4class{Compound}}, \code{\linkS4class{CompoundList}}, 
\code{\linkS4class{DRE}}, \code{\linkS4class{DREList}}, 
\code{\linkS4class{SynergyExperiment}}, \code{\linkS4class{SynergyExperimentList}}, \code{\linkS4class{SynergyData}}
}

