% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transferTrueValues.R
\docType{methods}
\name{transferTrueValues}
\alias{transferTrueValues}
\alias{transferTrueValues,Compound,Compound-method}
\alias{transferTrueValues,DRE,DRE-method}
\alias{transferTrueValues,SynergyScreen,SynergyScreen-method}
\title{Transfer "true" values of various parameters from one object to another}
\usage{
transferTrueValues(from, to)

\S4method{transferTrueValues}{Compound,Compound}(from, to)

\S4method{transferTrueValues}{DRE,DRE}(from, to)

\S4method{transferTrueValues}{SynergyScreen,SynergyScreen}(from, to)
}
\arguments{
\item{from}{object to transfer values from}

\item{to}{object to transfer values to}
}
\value{
Object \code{"to"} with modified slots

##  Stage 1: simulate individual compounds and their dose-response experiments
# Load a set of compounds and initialize a SynergyScreen object
cpds = readCompoundFile(system.file("extdata/15_cpds_simulation/compounds_1.csv",
                                    package="SynergyScreen"))
screen1 = new("SynergyScreen",compound_list=cpds)

#  Generate screen design, only testing single compounds at first
screen1 = generateDesign(screen1,pairs=FALSE)

#  Simulate compound growth curve characteristics and raw dose-response data
set.seed(20141120)
screen1 = simulate(screen1)

#  Normalize screen data, adjusting for plate bias
screen1 = normalize(screen1)

#  Check ranges of dose-response experiments
screen1 = checkRange(screen1)

## Stage 2: re-do dose-response experiments for which dose range checks failed, 
## with corrected dose ranges
#  Load the new set of compounds, initialize a SynergyScreen object and generate a design
cpds2 = readCompoundFile(system.file("extdata/15_cpds_simulation/compounds_2.csv",
                                     package="SynergyScreen"))
screen2 = new("SynergyScreen",compound_list=cpds2)
screen2 = generateDesign(screen2,pairs=FALSE)

#====================================================================================#
### Transfer previously simulated "true" IC50 and m values to compounds in screen2 ###
screen2 = transferTrueValues(screen1,screen2)
compound(screen1,"Cpd1")
compound(screen2,"Cpd1")
#====================================================================================##

#  Simulate dose-response data again
set.seed(20141121)
screen2 = simulate(screen2)
screen2 = normalize(screen2)
screen2 = checkRange(screen2)
}
\description{
Transfer "true" values of various parameters from one object to another
}
\details{
These methods are useful in multi-step simulations, when "ground truth" values, e.g. "true" values of \emph{IC50} and \emph{m} parameters of 
dose-response curves, need to be transferred from one step to the next
}
\section{Methods (by class)}{
\itemize{
\item \code{from = Compound,to = Compound}: Transfer "true" \emph{IC50} and \emph{m} values between simulated \code{Compound} objects

\item \code{from = DRE,to = DRE}: Transfer previously simulated "true" \emph{IC50} and \emph{m} values from one DRE object to another

\item \code{from = SynergyScreen,to = SynergyScreen}: Transfer previously simulated "true" \emph{IC50} and \emph{m} values from one SynergyScreen object to another
}}
\seealso{
\code{\linkS4class{Compound}}, \code{\linkS4class{DRE}}, \code{\linkS4class{SynergyScreen}},
}

